First cut at BitcoinCurrencyProvider
I started with the sample code from the Moneta User’s Guide
(which I think is out of date) I’m still learning the SPI so I may
be doing something seriously wrong.

I wrote 5 tests that are passing. Feel free to change the tests
or add some to show what is missing.